requirements.txt
gql[all]
django-crontab
graphene-django
